pyproject.toml
[tool.poetry]
name = "full-stack-ibm-course"
version = "0.1.0"
description = "Full Stack Application Development Project - IBM Course"
authors = ["Vinícius José Pierri Nogueira <vinijosenog@hotmail.com>"]
readme = "README.md"
packages = [{include = "full_stack_ibm_course"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
django = "3.1.3"
gunicorn = "20.1.0"
ibm-cloud-sdk-core = "3.10.0"
ibm-watson = "5.2.2"
ibmcloudant = "0.0.34"
pillow = "^10.0.0"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
pylint = "^2.17.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
